rest/src/requirements.txt
pyramid==1.10
requests==2.18
mysql-connector-python